+++ packages/orq-rc/src/orq_ai_sdk/models/createdatasourceop.py
@@ -314,7 +314,7 @@ class CreateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNVQVBAPGWTX3H2N3P12WQS3"
+        "01KNVVT6J4CWNVC2QAH6HBTVZ7"
     )
     r"""The unique identifier of the data source"""