+++ packages/orq-rc/src/orq_ai_sdk/models/listdatasourcesop.py
@@ -150,7 +150,7 @@ class ListDatasourcesData(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNVQVBAK6CHFKVE3RXH1CJFV"
+        "01KNVVT6J1490Q68467N6GTKT6"
     )
     r"""The unique identifier of the data source"""