+++ packages/orq-rc/src/orq_ai_sdk/models/updatedatasourceop.py
@@ -104,7 +104,7 @@ class UpdateDatasourceResponseBody(BaseModel):
     r"""The number of chunks in the datasource"""
 
     id: Annotated[Optional[str], pydantic.Field(alias="_id")] = (
-        "01KNVQVBARSMD1FF071MN7EA8R"
+        "01KNVVT6JD90HCSTYN0QDNQZ4W"
     )
     r"""The unique identifier of the data source"""